+++ engine/src/execution_graph/logic_controllers/CacheData.cpp
@@ -113,16 +113,16 @@ CPUCacheData::CPUCacheData(std::unique_ptr<ral::frame::BlazingTable> gpu_table, 
 }
 
 CPUCacheData::CPUCacheData(std::unique_ptr<ral::frame::BlazingTable> gpu_table,const MetadataDictionary & metadata, bool use_pinned)
-	: CacheData(CacheDataType::CPU, gpu_table->names(), gpu_table->get_schema(), gpu_table->num_rows()),
-	metadata(metadata)
+	: CacheData(CacheDataType::CPU, gpu_table->names(), gpu_table->get_schema(), gpu_table->num_rows())
 {
 	this->host_table = ral::communication::messages::serialize_gpu_message_to_host_table(gpu_table->toBlazingTableView(), use_pinned);
+	this->metadata = metadata;
 }
 
 CPUCacheData::CPUCacheData(const std::vector<blazingdb::transport::ColumnTransport> & column_transports,
 			std::vector<ral::memory::blazing_chunked_column_info> && chunked_column_infos,
 			std::vector<std::unique_ptr<ral::memory::blazing_allocation_chunk>> && allocations,
-			const MetadataDictionary & metadata) : metadata(metadata) {
+			const MetadataDictionary & metadata)  {
 
 	
 	this->cache_type = CacheDataType::CPU;
@@ -136,6 +136,7 @@ CPUCacheData::CPUCacheData(const std::vector<blazingdb::transport::ColumnTranspo
 		this->n_rows = column_transports[0].metadata.size;
 	}
 	this->host_table = std::make_unique<ral::frame::BlazingHostTable>(column_transports,std::move(chunked_column_infos), std::move(allocations));
+	this->metadata = metadata;
 }
 
 CPUCacheData::CPUCacheData(std::unique_ptr<ral::frame::BlazingHostTable> host_table)