+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 ## Bug Fixes
 - #1335 Fixing uninitialized var in orc metadata and handling the parseMetadata exceptions properly
 - #1339 Handling properly the nulls in case conditions with strings
+- #1348 Capturing error messages due to exceptions properly
 
 
 # BlazingSQL 0.18.0 (Date TBS)
@@ -33,6 +34,8 @@
 - #1284 Initial support for Windows Function
 - #1303 Add support for INITCAP
 - #1313 getting and using ORC metadata
+- #1347 Fixing issue when reading orc metadata from DATE dtype
+- #1338 Window Function support for LEAD and LAG statements 
 
 
 ## Improvements